fix: moar tests converted
